refactor: 答案アップロードテーブル機能の大規模リファクタリング
## 主要な改善点

### 📊 ファイル分割実績
- answer-sheet-table.tsx: 557行 → 96行 (82%削減)
- use-drag-drop.ts: 450行 → 139行 (69%削減)
- use-table-data.ts: 444行 → 131行 (70%削減)

### 🏗️ 新しい構造
- **型定義**: 機能別型定義の分離 (local-types.ts, drag-drop-types.ts)
- **コンポーネント**: 責任別分割 (table-content.tsx, table-drag-overlay.tsx)
- **フック**: 機能別分割 (状態管理, イベントハンドラー, データ生成)
- **ユーティリティ**: 共通関数の分離 (drag-drop-utils.ts, table-data-utils.ts)

### ✅ 品質向上
- 1ファイル1責任の原則を徹底
- 関連機能の論理的グループ化
- 型安全性の強化
- 保守性と開発効率の大幅向上

全ファイルが350行以下、ほとんどが200行以下の理想的な構造を実現。
