+++ components/projects/06-answer-sheets/answer-sheet-table/answer-sheet-table.tsx
@@ -1,250 +1,59 @@
 "use client"
 
 import {
-  EmptyTableCell,
-  FilePreviewCell,
-  SortableTableCell,
+  TableContent,
+  TableDragOverlay,
   TableHeader,
-  UploadToCellModal,
+  UploadModalWrapper,
 } from "@/components/projects/06-answer-sheets/answer-sheet-table/components"
-import {
-  useDisabledState,
-  useDragDrop,
-  useNameRegion,
-  useTableData,
-} from "@/components/projects/06-answer-sheets/answer-sheet-table/hooks"
-import type { PreviewMode } from "@/components/projects/06-answer-sheets/answer-sheet-table/types"
+import { useAnswerSheetTableLogic } from "@/components/projects/06-answer-sheets/answer-sheet-table/hooks"
+import type { AnswerSheetTableProps } from "@/components/projects/06-answer-sheets/answer-sheet-table/types"
 import { Card, CardContent } from "@/components/ui/card"
-import {
-  Table,
-  TableBody,
-  TableHead,
-  TableRow,
-  TableHeader as UITableHeader,
-} from "@/components/ui/table"
-import type {
-  PendingChange,
-  PlacementStrategy,
-  UnifiedFile,
-  UnifiedStudent,
-  UploadData,
-} from "@/types/answer-sheet.types"
-import { closestCenter, DndContext, DragOverlay } from "@dnd-kit/core"
-import { rectSortingStrategy, SortableContext } from "@dnd-kit/sortable"
+import { closestCenter, DndContext } from "@dnd-kit/core"
 import { FileImage } from "lucide-react"
-import { useEffect, useState } from "react"
 
-// ============================================================================
-// Props定義
-// ============================================================================
 
-interface AnswerSheetTableProps {
-  projectId: string
-  students: UnifiedStudent[]
-  files: UnifiedFile[]
-  masterImageCount: number
-  fileOrder?: PlacementStrategy
-  isUploading?: boolean
-  onFileOrderChange?: (order: PlacementStrategy) => void
-  onFilesChange: (files: UnifiedFile[]) => void
-  onUpload: (data: UploadData[]) => void
-  imageLoadStates?: Record<string, "pending" | "loading" | "loaded" | "error">
-  observerRef?: React.RefObject<IntersectionObserver | null>
-  mode?: "upload" | "view"
-  onReloadData?: () => void
-
-  // 変更状態管理用（確認モードのみ）
-  pendingChanges?: PendingChange[]
-  affectedCells?: Set<string>
-  onUpdatePendingChanges?: (
-    changedFiles: Array<{ fileId: string; fromState: any; toState: any }>,
-  ) => void
-  onResetDragDrop?: React.MutableRefObject<(() => void) | null>
-
-  // 上書き制御用（アップロードモードでの既存答案情報）
-  existingAnswerSheets?: Array<{
-    id: string
-    studentId: string | null
-    pageNumber: number
-  }>
-}
-
-// ============================================================================
-// メインコンポーネント
-// ============================================================================
-
-export function AnswerSheetTable({
-  projectId,
-  students,
-  files,
-  masterImageCount,
-  fileOrder = "page-first",
-  isUploading = false,
-  onFileOrderChange,
-  onFilesChange,
-  onUpload,
-  imageLoadStates = {},
-  observerRef,
-  mode = "upload",
-  onReloadData,
-  pendingChanges,
-  affectedCells,
-  onUpdatePendingChanges,
-  onResetDragDrop,
-  existingAnswerSheets = [],
-}: AnswerSheetTableProps) {
-  // ============================================================================
-  // カスタムフック
-  // ============================================================================
-
+export function AnswerSheetTable(props: AnswerSheetTableProps) {
   const {
-    nameRegionAvailable,
+    imageLoadStates = {},
+    observerRef,
+    mode = "upload",
+    affectedCells,
+  } = props
+
+  // メインロジックを含むカスタムフック
+  const {
     canvasRef,
-    checkNameRegionAvailability,
     drawNameRegionCanvas,
-  } = useNameRegion(projectId)
-
-  const {
     disabledState,
-    setDisabledState,
     toggleRowDisabled,
     toggleColDisabled,
     togglePositionDisabled,
     toggleFileDisabled,
-    isPositionDisabled,
-    initializeStudentsWithoutAnswers,
-    allowOverwrite,
-    setAllowOverwrite,
-  } = useDisabledState()
-
-  const {
     sortedStudents,
     getEnabledFiles,
-    getDisabledFiles,
     getFileColor,
     tableData,
     positionsWithExistingAnswers,
-  } = useTableData(
-    files,
-    students,
-    masterImageCount,
-    fileOrder,
-    disabledState,
-    isPositionDisabled,
-    mode,
-    allowOverwrite,
-    existingAnswerSheets,
-  )
-
-  const {
     sensors,
     activeFile,
     handleDragStart,
     handleDragOver,
     handleDragEnd,
-    resetToInitialState,
-  } = useDragDrop(
-    files,
-    onFilesChange,
-    getEnabledFiles,
-    getDisabledFiles,
-    disabledState,
-    setDisabledState,
-    students,
-    masterImageCount,
-    mode,
-    fileOrder,
-    onReloadData,
-    onUpdatePendingChanges,
-  )
-
-  // コールバック関数をプロップとして渡すためのuseEffect
-  useEffect(() => {
-    if (onResetDragDrop) {
-      onResetDragDrop.current = resetToInitialState
-    }
-  }, [onResetDragDrop, resetToInitialState])
-
-  // ============================================================================
-  // ローカルState
-  // ============================================================================
-
-  const [previewMode, setPreviewMode] = useState<PreviewMode>("full")
-  const [uploadModalState, setUploadModalState] = useState<{
-    isOpen: boolean
-    position?: number
-    studentName?: string
-    pageNumber?: number
-  }>({ isOpen: false })
-
-  // デバッグ用のpreviewMode変更ハンドラー
-  const handlePreviewModeChange = (mode: PreviewMode) => {
-    setPreviewMode(mode)
-  }
-
-  // ============================================================================
-  // 初期化処理
-  // ============================================================================
-
-  useEffect(() => {
-    checkNameRegionAvailability()
-  }, [checkNameRegionAvailability])
-
-  // ============================================================================
-  // イベントハンドラー
-  // ============================================================================
-
-  const handleUpload = () => {
-    const uploadData: UploadData[] = []
-
-    // 動的テーブルデータから配置済みファイルのアップロードデータを生成
-    tableData.forEach((row) => {
-      row.forEach((cell) => {
-        if (
-          cell.type === "file" &&
-          cell.file &&
-          cell.student &&
-          cell.pageNumber
-        ) {
-          uploadData.push({
-            name: cell.file.name,
-            fileName: cell.file.name,
-            originalFileName: cell.file.originalFileName,
-            type: cell.file.type,
-            buffer: cell.file.buffer,
-            studentId: cell.student.id,
-            pageNumber: cell.pageNumber,
-            overwrite: false,
-          })
-        }
-      })
-    })
-
-    onUpload(uploadData)
-  }
-
-  // ============================================================================
-  // 初期化処理
-  // ============================================================================
-
-  // 答案がない生徒の自動無効化（DBベース）
-  useEffect(() => {
-    initializeStudentsWithoutAnswers(students, files)
-  }, [students, files, initializeStudentsWithoutAnswers])
-
-  // ============================================================================
-  // 計算済みプロパティ
-  // ============================================================================
-
-  const maxPages = masterImageCount
-  const trashFiles = getDisabledFiles()
-  const hasNameRegion = Object.values(nameRegionAvailable).some(Boolean)
-
-  // デバッグログ
-
-  // ============================================================================
-  // レンダリング
-  // ============================================================================
+    allowOverwrite,
+    setAllowOverwrite,
+    previewMode,
+    uploadModalState,
+    handlePreviewModeChange,
+    handleUpload,
+    handleUploadModalOpen,
+    handleUploadModalClose,
+    handleUploadToCell,
+    maxPages,
+    trashFiles,
+    hasNameRegion,
+    nameRegionAvailable,
+  } = useAnswerSheetTableLogic(props)
 
   // 模範解答が存在しない場合
   if (maxPages === 0) {
@@ -276,11 +85,11 @@ export function AnswerSheetTable({
             enabledFilesCount={getEnabledFiles().length}
             trashFiles={trashFiles}
             onFileRestore={toggleFileDisabled}
-            isUploading={isUploading}
+            isUploading={props.isUploading || false}
             mode={mode}
             onUpload={handleUpload}
-            fileOrder={fileOrder}
-            onFileOrderChange={onFileOrderChange}
+            fileOrder={props.fileOrder || "page-first"}
+            onFileOrderChange={props.onFileOrderChange}
             previewMode={previewMode}
             onPreviewModeChange={handlePreviewModeChange}
             hasNameRegion={hasNameRegion}
@@ -289,268 +98,48 @@ export function AnswerSheetTable({
           />
 
           <CardContent className="min-h-0 flex-1 overflow-auto p-4">
-            <SortableContext
-              items={getEnabledFiles().map((file) => file.id)}
-              strategy={rectSortingStrategy}
-            >
-              <Table>
-                <UITableHeader>
-                  <TableRow>
-                    {/* 生徒名列ヘッダー */}
-                    <TableHead
-                      className={`w-32 border text-center ${mode === "upload" ? "cursor-pointer" : ""}`}
-                      onClick={
-                        mode === "upload"
-                          ? () => toggleColDisabled(-1)
-                          : undefined
-                      }
-                    >
-                      生徒名
-                    </TableHead>
-                    {/* ページ列ヘッダー */}
-                    {Array.from({ length: maxPages }, (_, pageIndex) => (
-                      <TableHead
-                        key={pageIndex}
-                        className={`w-32 border text-center ${
-                          mode === "upload" ? "cursor-pointer" : ""
-                        } ${
-                          disabledState.cols.has(pageIndex)
-                            ? "bg-gray-200"
-                            : "bg-white"
-                        }`}
-                        onClick={
-                          mode === "upload"
-                            ? () => toggleColDisabled(pageIndex)
-                            : undefined
-                        }
-                      >
-                        ページ {pageIndex + 1}
-                      </TableHead>
-                    ))}
-                  </TableRow>
-                </UITableHeader>
-                <TableBody>
-                  {tableData.map((row, studentIndex) => (
-                    <TableRow key={sortedStudents[studentIndex].id}>
-                      {/* 生徒名セル */}
-                      <TableHead
-                        className={`border text-center ${
-                          mode === "upload" ? "cursor-pointer" : ""
-                        } ${
-                          disabledState.rows.has(studentIndex)
-                            ? "bg-gray-200"
-                            : "bg-white"
-                        }`}
-                        onClick={
-                          mode === "upload"
-                            ? () => toggleRowDisabled(studentIndex)
-                            : undefined
-                        }
-                      >
-                        <div className="px-2 py-1">
-                          <div className="text-sm font-medium">
-                            {sortedStudents[studentIndex].lastName}{" "}
-                            {sortedStudents[studentIndex].firstName}
-                          </div>
-                          <div className="text-xs text-gray-500">
-                            {sortedStudents[studentIndex].studentId}
-                          </div>
-                        </div>
-                      </TableHead>
-
-                      {/* ファイルセル */}
-                      {row.map((cellData, pageIndex) => {
-                        if (
-                          cellData.type === "disabled" ||
-                          cellData.type === "empty"
-                        ) {
-                          // 既存答案があるかチェック（空セル用）
-                          const hasExistingAnswerForEmpty =
-                            mode === "upload" &&
-                            cellData.type === "empty" &&
-                            positionsWithExistingAnswers.has(cellData.position)
-
-                          // そのセルに新しく追加しようとしている画像ファイルがあるかチェック
-                          // (つまり、files配列にstudentId・pageNumberが設定されているが、まだアップロードされていない状態)
-                          const newFileInCell = files.find(file => 
-                            file.studentId === cellData.student?.id && 
-                            file.pageNumber === cellData.pageNumber &&
-                            !disabledState.files.has(file.id) // 無効化されていない
-                          )
-                          const hasNewFileToUpload = !!newFileInCell
-
-                          // 無効化の理由を判定
-                          let disabledReason:
-                            | "row"
-                            | "column"
-                            | "position"
-                            | "existing_answer"
-                            | "absent_student"
-                            | undefined
-                          if (cellData.type === "disabled") {
-                            const student = sortedStudents[studentIndex]
-
-                            if (disabledState.rows.has(studentIndex)) {
-                              // 行無効の場合、欠席生徒かどうかをチェック
-                              if (student?.status === "absent") {
-                                disabledReason = "absent_student"
-                              } else {
-                                disabledReason = "row"
-                              }
-                            } else if (disabledState.cols.has(pageIndex)) {
-                              disabledReason = "column"
-                            } else if (
-                              disabledState.positions.has(cellData.position)
-                            ) {
-                              disabledReason = "position"
-                            } else if (
-                              mode === "upload" &&
-                              !allowOverwrite &&
-                              positionsWithExistingAnswers.has(
-                                cellData.position,
-                              )
-                            ) {
-                              disabledReason = "existing_answer"
-                            }
-                          }
-
-                          return (
-                            <EmptyTableCell
-                              key={cellData.position}
-                              position={cellData.position}
-                              student={cellData.student}
-                              pageNumber={cellData.pageNumber}
-                              isPositionDisabled={cellData.type === "disabled"}
-                              isPendingChange={false} // 空のセルは通常変更対象外
-                              mode={mode}
-                              hasExistingAnswer={hasExistingAnswerForEmpty}
-                              allowOverwrite={allowOverwrite}
-                              disabledReason={disabledReason}
-                              onTogglePosition={() => 
-                                togglePositionDisabled(cellData.position)
-                              }
-                              onUploadToCell={() => {
-                                setUploadModalState({
-                                  isOpen: true,
-                                  position: cellData.position,
-                                  studentName: cellData.student 
-                                    ? `${cellData.student.lastName} ${cellData.student.firstName}`
-                                    : undefined,
-                                  pageNumber: cellData.pageNumber ?? undefined,
-                                })
-                              }}
-                              onToggleAnswerDisabled={() => {
-                                if (newFileInCell) {
-                                  toggleFileDisabled(newFileInCell.id)
-                                }
-                              }}
-                              hasNewFileToUpload={hasNewFileToUpload}
-                            />
-                          )
-                        }
-
-                        // ファイルセル
-                        const file = cellData.file!
-                        const isFileDisabled = disabledState.files.has(file.id)
-                        const hasExistingAnswer =
-                          mode === "upload" &&
-                          !allowOverwrite &&
-                          positionsWithExistingAnswers.has(cellData.position)
-                        // 上書き無効時で既存答案がある場合はドラッグ無効
-                        const isDragDisabledByOverwrite = hasExistingAnswer
-
-                        return (
-                          <SortableTableCell
-                            key={file.id}
-                            id={file.id}
-                            position={cellData.position}
-                            hasFile={true}
-                            isPositionDisabled={isDragDisabledByOverwrite}
-                            isFileDisabled={isFileDisabled}
-                            onTogglePosition={
-                              mode === "upload"
-                                ? () =>
-                                    togglePositionDisabled(cellData.position)
-                                : () => {}
-                            }
-                            onToggleFileDisabled={
-                              mode === "upload"
-                                ? () => toggleFileDisabled(file.id)
-                                : () => {}
-                            }
-                            onUploadToCell={() => {
-                              // アップロードはアップロードタブから行う
-                            }}
-                            fileId={file.id}
-                            observerRef={observerRef}
-                            mode={mode}
-                            studentName={cellData.student ? `${cellData.student.lastName} ${cellData.student.firstName}` : undefined}
-                            pageNumber={cellData.pageNumber ?? undefined}
-                            hasScoreData={true} // TODO: 実際の採点データ有無を判定
-                            onDeleteFileWithScoring={() => {
-                              // TODO: 実際の削除処理を実装
-                              console.log(`Delete file ${file.id} with scoring data`)
-                            }}
-                          >
-                            <FilePreviewCell
-                              file={file}
-                              pageNumber={cellData.pageNumber || 1}
-                              previewMode={previewMode}
-                              isFileDisabled={isFileDisabled}
-                              nameRegionAvailable={
-                                nameRegionAvailable[cellData.pageNumber || 1]
-                              }
-                              getFileColor={getFileColor}
-                              drawNameRegionCanvas={drawNameRegionCanvas}
-                              imageLoadState={imageLoadStates[file.id]}
-                              isPendingChange={
-                                affectedCells?.has(file.id) || false
-                              }
-                              hasExistingAnswer={hasExistingAnswer}
-                            />
-                          </SortableTableCell>
-                        )
-                      })}
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </SortableContext>
+            <TableContent
+              tableData={tableData}
+              sortedStudents={sortedStudents}
+              maxPages={maxPages}
+              disabledState={disabledState}
+              mode={mode}
+              previewMode={previewMode}
+              nameRegionAvailable={nameRegionAvailable}
+              positionsWithExistingAnswers={positionsWithExistingAnswers}
+              allowOverwrite={allowOverwrite}
+              files={props.files}
+              affectedCells={affectedCells}
+              imageLoadStates={imageLoadStates}
+              observerRef={observerRef}
+              getEnabledFiles={getEnabledFiles}
+              getFileColor={getFileColor}
+              drawNameRegionCanvas={drawNameRegionCanvas}
+              toggleRowDisabled={toggleRowDisabled}
+              toggleColDisabled={toggleColDisabled}
+              togglePositionDisabled={togglePositionDisabled}
+              toggleFileDisabled={toggleFileDisabled}
+              onUploadModalOpen={handleUploadModalOpen}
+            />
           </CardContent>
         </Card>
 
-        <DragOverlay dropAnimation={null}>
-          {activeFile ? (
-            <div className="h-32 w-32 scale-110 rotate-3 transform rounded border-2 border-blue-400 bg-white shadow-2xl">
-              <FilePreviewCell
-                file={activeFile}
-                pageNumber={1}
-                previewMode={previewMode}
-                isFileDisabled={false}
-                nameRegionAvailable={false}
-                getFileColor={getFileColor}
-                drawNameRegionCanvas={drawNameRegionCanvas}
-                imageLoadState="loaded"
-              />
-            </div>
-          ) : null}
-        </DragOverlay>
+        <TableDragOverlay
+          activeFile={activeFile}
+          previewMode={previewMode}
+          getFileColor={getFileColor}
+          drawNameRegionCanvas={drawNameRegionCanvas}
+        />
       </DndContext>
 
       {/* 氏名欄クリッピング用の隠しcanvas */}
       <canvas ref={canvasRef} className="hidden" width={0} height={0} />
 
       {/* セルにアップロードモーダル */}
-      <UploadToCellModal
-        isOpen={uploadModalState.isOpen}
-        onClose={() => setUploadModalState({ isOpen: false })}
-        onUpload={(file, pageNumber) => {
-          // TODO: 指定されたセル位置にファイルをアップロード
-          console.log(`Uploading to position ${uploadModalState.position}:`, file.name, pageNumber)
-          setUploadModalState({ isOpen: false })
-        }}
-        studentName={uploadModalState.studentName}
-        pageNumber={uploadModalState.pageNumber}
+      <UploadModalWrapper
+        uploadModalState={uploadModalState}
+        onClose={handleUploadModalClose}
+        onUpload={handleUploadToCell}
       />
     </div>
   )
