+++ components/projects/06-answer-sheets/answer-sheet-table/components/index.ts
@@ -5,5 +5,8 @@ export { OverwriteToggle } from "./overwrite-toggle"
 export { PlacementStrategySelector } from "./placement-strategy-selector"
 export { PreviewModeToggle } from "./preview-mode-toggle"
 export { SortableTableCell } from "./sortable-table-cell"
+export { TableContent } from "./table-content"
+export { TableDragOverlay } from "./table-drag-overlay"
 export { TableHeader } from "./table-header"
+export { UploadModalWrapper } from "./upload-modal-wrapper"
 export { UploadToCellModal } from "./upload-to-cell-modal"
