+++ components/projects/06-answer-sheets/answer-sheet-table/types/index.ts
@@ -4,6 +4,10 @@ import type {
   UnifiedStudent,
 } from "@/types/answer-sheet.types"
 
+// Re-export local types
+export type { AnswerSheetTableProps, DisabledReason, UploadModalState } from "./local-types"
+export type { FileState, UseDragDropParams, UseDragDropReturn } from "./drag-drop-types"
+
 // Preview mode for different display options
 export type PreviewMode = "full" | "name-only"
 
@@ -19,8 +23,8 @@ export interface ExtendedDisabledState {
 export interface CellData {
   type: "file" | "empty" | "disabled"
   position: number
-  student: UnifiedStudent | null
-  pageNumber: number | null
+  student?: UnifiedStudent
+  pageNumber?: number
   file?: UnifiedFile
 }
