+++ components/projects/06-answer-sheets/answer-sheet-table/utils/index.ts
@@ -1,0 +1,14 @@
+export {
+  buildDnDArrayFromFileStates,
+  compareFileStates,
+  updateFileStatesFromDnDArray,
+} from "./drag-drop-utils"
+
+export {
+  calculateDynamicDisabledPositions,
+  calculatePositionsWithExistingAnswers,
+  getDisabledFiles,
+  getEnabledFiles,
+  getFileColor,
+  sortStudentsByCustomOrder,
+} from "./table-data-utils"
